+++ packages/model-runtime/src/index.ts
@@ -1,4 +1,4 @@
-export * from './BaseAI';
+export * from './core/BaseAI';
 export { ModelRuntime } from './core/ModelRuntime';
 export { createOpenAICompatibleRuntime } from './core/openaiCompatibleFactory';
 export * from './helpers';
