+++ packages/model-runtime/src/providers/bfl/index.ts
@@ -1,7 +1,7 @@
 import createDebug from 'debug';
 import { ClientOptions } from 'openai';
 
-import { LobeRuntimeAI } from '../../BaseAI';
+import { LobeRuntimeAI } from '../../core/BaseAI';
 import { AgentRuntimeErrorType } from '../../types/error';
 import { CreateImagePayload, CreateImageResponse } from '../../types/image';
 import { AgentRuntimeError } from '../../utils/createError';
